wip: Migrate plugin functions.
